+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### SNAPSHOT
+* Added ability to automatically start new downloads
 * Added ability to restart download from the beginning
 * Added tray icon and minimize to tray setting
 * Added download speed and remaining time columns, implemented related tracking and calculations